[FEATURE] Add PanelActions component with header icon buttons for panel interactions
