+++ ui/dashboards/src/components/Panel/HeaderIconButton.tsx
@@ -1,0 +1,6 @@
+import { IconButton, styled } from '@mui/material';
+
+export const HeaderIconButton = styled(IconButton)(({ theme }) => ({
+  borderRadius: theme.shape.borderRadius,
+  padding: '4px',
+}));
