+++ ui/dashboards/src/components/Panel/PanelHeader.tsx
@@ -1,4 +1,4 @@
-// Copyright 2023 The Perses Authors
+// Copyright 2025 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,62 +11,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { CardHeader, Typography, Stack, CardHeaderProps, styled, IconButton, Box } from '@mui/material';
+import { CardHeader, CardHeaderProps, Stack, Typography } from '@mui/material';
 import { InfoTooltip, combineSx } from '@perses-dev/components';
-import InformationOutlineIcon from 'mdi-material-ui/InformationOutline';
-import PencilIcon from 'mdi-material-ui/PencilOutline';
-import DeleteIcon from 'mdi-material-ui/DeleteOutline';
-import DragIcon from 'mdi-material-ui/DragVertical';
-import ArrowExpandIcon from 'mdi-material-ui/ArrowExpand';
-import ArrowCollapseIcon from 'mdi-material-ui/ArrowCollapse';
-import ContentCopyIcon from 'mdi-material-ui/ContentCopy';
-import { useReplaceVariablesInString } from '@perses-dev/plugin-system';
-import { PropsWithChildren, ReactElement, ReactNode, useState } from 'react';
 import { Link } from '@perses-dev/core';
-import MenuIcon from 'mdi-material-ui/Menu';
-import Popover from '@mui/material/Popover';
-import { ARIA_LABEL_TEXT, TOOLTIP_TEXT } from '../../constants';
+import { useReplaceVariablesInString } from '@perses-dev/plugin-system';
+import InformationOutlineIcon from 'mdi-material-ui/InformationOutline';
+import { ReactElement, ReactNode } from 'react';
+import { HEADER_ACTIONS_CONTAINER_NAME } from '../../constants';
+import { PanelActions, PanelActionsProps } from './PanelActions';
 import { PanelLinks } from './PanelLinks';
-
-const ShowAction: React.FC<PropsWithChildren> = ({ children }) => {
-  const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
-
-  const handleClick = (event: React.MouseEvent<HTMLElement>): undefined => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = (): undefined => {
-    setAnchorEl(null);
-  };
-
-  const open = Boolean(anchorEl);
-  const id = open ? 'simple-popover' : undefined;
-
-  return (
-    <>
-      <HeaderIconButton
-        aria-describedby={id}
-        onClick={handleClick}
-        aria-label={ARIA_LABEL_TEXT.editPanel('expand')}
-        size="small"
-      >
-        <MenuIcon fontSize="inherit" />
-      </HeaderIconButton>
-      <Popover
-        id={id}
-        open={open}
-        anchorEl={anchorEl}
-        onClose={handleClose}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'left',
-        }}
-      >
-        <Box sx={{ padding: '8px' }}>{children}</Box>
-      </Popover>
-    </>
-  );
-};
+import { HeaderIconButton } from './HeaderIconButton';
 
 type OmittedProps = 'children' | 'action' | 'title' | 'disableTypography';
 
@@ -76,15 +30,8 @@ export interface PanelHeaderProps extends Omit<CardHeaderProps, OmittedProps> {
   description?: string;
   links?: Link[];
   extra?: ReactNode;
-  readHandlers?: {
-    isPanelViewed?: boolean;
-    onViewPanelClick: () => void;
-  };
-  editHandlers?: {
-    onEditPanelClick: () => void;
-    onDuplicatePanelClick: () => void;
-    onDeletePanelClick: () => void;
-  };
+  readHandlers?: PanelActionsProps['readHandlers'];
+  editHandlers?: PanelActionsProps['editHandlers'];
 }
 
 export function PanelHeader({
@@ -103,67 +50,6 @@ export function PanelHeader({
 
   const title = useReplaceVariablesInString(rawTitle) as string;
   const description = useReplaceVariablesInString(rawDescription);
-
-  let readActions: CardHeaderProps['action'] = undefined;
-  if (readHandlers !== undefined) {
-    readActions = (
-      <InfoTooltip description={TOOLTIP_TEXT.viewPanel}>
-        <HeaderIconButton
-          aria-label={ARIA_LABEL_TEXT.viewPanel(title)}
-          size="small"
-          onClick={readHandlers.onViewPanelClick}
-        >
-          {readHandlers.isPanelViewed ? (
-            <ArrowCollapseIcon fontSize="inherit" />
-          ) : (
-            <ArrowExpandIcon fontSize="inherit" />
-          )}
-        </HeaderIconButton>
-      </InfoTooltip>
-    );
-  }
-  let editActions: CardHeaderProps['action'] = undefined;
-  if (editHandlers !== undefined) {
-    // If there are edit handlers, always just show the edit buttons
-    editActions = (
-      <>
-        <InfoTooltip description={TOOLTIP_TEXT.editPanel}>
-          <HeaderIconButton
-            aria-label={ARIA_LABEL_TEXT.editPanel(title)}
-            size="small"
-            onClick={editHandlers.onEditPanelClick}
-          >
-            <PencilIcon fontSize="inherit" />
-          </HeaderIconButton>
-        </InfoTooltip>
-        <InfoTooltip description={TOOLTIP_TEXT.duplicatePanel}>
-          <HeaderIconButton
-            aria-label={ARIA_LABEL_TEXT.duplicatePanel(title)}
-            size="small"
-            onClick={editHandlers.onDuplicatePanelClick}
-          >
-            <ContentCopyIcon
-              fontSize="inherit"
-              sx={{
-                // Shrink this icon a little bit to look more consistent
-                // with the other icons in the header.
-                transform: 'scale(0.925)',
-              }}
-            />
-          </HeaderIconButton>
-        </InfoTooltip>
-        <InfoTooltip description={TOOLTIP_TEXT.deletePanel}>
-          <HeaderIconButton
-            aria-label={ARIA_LABEL_TEXT.deletePanel(title)}
-            size="small"
-            onClick={editHandlers.onDeletePanelClick}
-          >
-            <DeleteIcon fontSize="inherit" />
-          </HeaderIconButton>
-        </InfoTooltip>
-      </>
-    );
-  }
 
   return (
     <CardHeader
@@ -205,51 +91,11 @@ export function PanelHeader({
           {links !== undefined && links.length > 0 && <PanelLinks links={links} />}
         </Stack>
       }
-      action={
-        <HeaderActionWrapper
-          sx={combineSx((theme) => {
-            theme.containerQueries('primary').up(200);
-            return {};
-          })}
-          direction="row"
-          spacing={0.25}
-          alignItems="center"
-        >
-          <Box
-            sx={{
-              '@container responsiveBox (min-width: 200px)': {
-                display: 'block',
-              },
-              display: 'none',
-            }}
-          >
-            {editHandlers === undefined && extra} {readActions} {editActions}
-          </Box>
-          {editActions && (
-            <>
-              <Box
-                sx={{
-                  '@container responsiveBox (max-width: 200px)': {
-                    display: 'block',
-                  },
-                  display: 'none',
-                }}
-              >
-                <ShowAction>{editActions}</ShowAction>
-              </Box>
-              <InfoTooltip description={TOOLTIP_TEXT.movePanel}>
-                <HeaderIconButton aria-label={ARIA_LABEL_TEXT.movePanel(title)} size="small">
-                  <DragIcon className="drag-handle" sx={{ cursor: 'grab' }} fontSize="inherit" />
-                </HeaderIconButton>
-              </InfoTooltip>
-            </>
-          )}
-        </HeaderActionWrapper>
-      }
+      action={<PanelActions title={title} readHandlers={readHandlers} editHandlers={editHandlers} extra={extra} />}
       sx={combineSx(
         (theme) => ({
           containerType: 'inline-size',
-          containerName: 'responsiveBox',
+          containerName: HEADER_ACTIONS_CONTAINER_NAME,
           padding: theme.spacing(1),
           borderBottom: `solid 1px ${theme.palette.divider}`,
           '.MuiCardHeader-content': {
@@ -269,17 +115,3 @@ export function PanelHeader({
     />
   );
 }
-
-const HeaderIconButton = styled(IconButton)(({ theme }) => ({
-  borderRadius: theme.shape.borderRadius,
-  padding: '4px',
-}));
-
-const HeaderActionWrapper = styled(Stack)(() => ({
-  // Adding back the negative margins from MUI's defaults for actions, so we
-  // avoid increasing the header size when actions are present while also being
-  // able to vertically center the actions.
-  // https://github.com/mui/material-ui/blob/master/packages/mui-material/src/CardHeader/CardHeader.js#L56-L58
-  marginTop: -4,
-  marginBottom: -4,
-}));
