+++ ui/dashboards/src/constants/styles.ts
@@ -23,3 +23,6 @@ export const editButtonStyle: SxProps<Theme> = {
 
 export const MIN_VARIABLE_WIDTH = 120;
 export const MAX_VARIABLE_WIDTH = 500;
+
+export const HEADER_ACTIONS_MIN_WIDTH = 200;
+export const HEADER_ACTIONS_CONTAINER_NAME = 'header-actions-container';
